Schlopey
The colour map is now a globally stored variable. This slightly speeds up generation, but it's main purpose is to make the colours easier to keep track of for brushing, etc.

Also when colouring, the shader now takes into account the slope of the terrain. This means cliffs no longer have awkard green bands on them just because they pass through the grass height.