Working demo for Brush
With a click, and only with a click, the terrain can be updated, either in a circle or a square. Terrain can be heightened or colour. Heightening works near chunk borders, but colouring doesn't.
There's no UI, it doesn't work on holding down the button, and colouring doesn't work too well yet. Also there's lag and the RayCast only collides with water, not terrain, because terrain has no collision.